backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.28
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
langchain>=0.1.0
langchain-community>=0.0.38
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.1
python-socketio==5.11.0
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
unstructured==0.11.8
python-magic==0.4.27
pdf2image==1.17.0
pytesseract==0.3.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
scikit-learn==1.4.1.post1
scipy==1.12.0
docx2txt==0.8
PyPDF2==3.0.1
aiofiles==23.2.1
openai==1.11.1
tiktoken==0.5.2 
pdfminer==20191125
pdfminer.six==20221105
opencv-python==4.9.0.80
python-pptx==1.0.2
